Add unit tests for shared submit utils
